renamed schema versions v1 and v2
