+++ client/schema/index.js
@@ -1,7 +1,7 @@
-import aspel from './aspel';
-import marvell from './marvell';
+import v1 from './v1';
+import v2 from './v2';
 
 export default {
-  0: aspel,
-  1: marvell
+  0: v1,
+  1: v2
 }
